Move the CopyDirectory() and RemoveDirectory() functions into FileUtils.cpp.
